requirements.txt
flask
flask-cors
Pillow
PyMuPDF
PyPDF2
pdf2image
fastapi
uvicorn
python-multipart
pydantic
